BC_TRY_RUN_OUTPUT wasn't set up to take no arguments.  fix that so it works with AC_REQUIRE.
svn:revision:32474
svn:branch:trunk
svn:account:brlcad

+++ m4/epsilon.m4
@@ -62,12 +62,15 @@ cat > conftest.$ac_ext <<EOF
 #include "confdefs.h"
 [$2]
 EOF
-if AC_TRY_EVAL(ac_link) && test -s conftest${ac_exeext} && (./conftest > conftest.stdout; exit) 2>/dev/null ; then
-	$1=`cat ./conftest.stdout`
-else
+__var=[$1]
+if test "x$__var" != "x" ; then
+    if AC_TRY_EVAL(ac_link) && test -s conftest${ac_exeext} && (./conftest > conftest.stdout; exit) 2>/dev/null ; then
+	[$1]=`cat ./conftest.stdout`
+    else
 	echo "configure: failed program was:" >&AC_FD_CC
 	cat conftest.$ac_ext >&AC_FD_CC
-	$1=""
+	[$1]=""
+    fi
 fi
 rm -fr conftest*
 ])
@@ -75,11 +78,14 @@ rm -fr conftest*
 
 # BC_FLOAT_EPSILON()
 AC_DEFUN([BC_FLOAT_EPSILON], [
+
+AC_REQUIRE([BC_TRY_RUN_OUTPUT])
+
 dnl determine the minimum single-precision floating point tolerance
 dnl value at compile time such that: 1.0 + value != 1.0
 dnl ANSI defines this as FLT_EPSILON but float.h may not provide it.
 AC_MSG_CHECKING([single-precision floating point tolerance])
-AC_REQUIRE([BC_TRY_RUN_OUTPUT])
+
 BC_TRY_RUN_OUTPUT(bc_flt_epsilon, [
 #include <stdio.h>
 int main (int ac, char *av[]) {
@@ -98,11 +104,14 @@ AC_MSG_RESULT([$bc_flt_epsilon])
 
 # BC_DOUBLE_EPSILON()
 AC_DEFUN([BC_DOUBLE_EPSILON], [
+
 AC_REQUIRE([BC_TRY_RUN_OUTPUT])
+
 dnl determine the minimum double-precision floating point tolerance
 dnl value at compile time such that: 1.0 + value != 1.0
 dnl ANSI defines this as DBL_EPSILON but float.h may not provide it.
 AC_MSG_CHECKING([double-precision floating point tolerance])
+
 BC_TRY_RUN_OUTPUT(bc_dbl_epsilon, [
 #include <stdio.h>
 int main (int ac, char *av[]) {
